requirements.txt
asgiref==3.5.2
Django==4.1.4
django-ckeditor==6.5.1
django-cuser==2017.3.16
django-js-asset==2.0.0
django-modeltranslation==0.18.7
django-phone-field==1.8.1
django-tinymce==3.5.0
Pillow==9.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
